Modularlize GQL request strings
